24-09-01 6h14

+++ Assets/Data/Script/Stat/Skill/Weapon/Shoot/Shotgun/ShotgunStat.cs
@@ -8,5 +8,5 @@ public class ShotgunStat : ShootSkillStat
 {
     [Header("Shotgun Stat")]
     public float SpreadRange;
-    public float PelletAmount;
+    public int PelletAmount;
 }
